+++ crates/crypto/src/objects/memory.rs
