+++ crates/crypto/src/objects/stream.rs
